Move sync_metadata related methods back to MetadataSyncUtils to avoid circular imports
