+++ anchore_engine/services/policy_engine/engine/feeds/sync_utils.py
@@ -1,7 +1,11 @@
 from abc import ABC, abstractmethod
 from typing import Dict, List, Optional, Tuple, Union
 
+from sqlalchemy.orm.session import Session
+
+from anchore_engine.common.schemas import FeedAPIGroupRecord, FeedAPIRecord
 from anchore_engine.db import FeedGroupMetadata, FeedMetadata
+from anchore_engine.db import get_thread_scoped_session as get_session
 from anchore_engine.services.policy_engine.engine.feeds import IFeedSource
 from anchore_engine.services.policy_engine.engine.feeds.client import (
     FeedServiceClient,
@@ -10,17 +14,221 @@ from anchore_engine.services.policy_engine.engine.feeds.client import (
     get_grype_db_client,
 )
 from anchore_engine.services.policy_engine.engine.feeds.config import SyncConfig
+from anchore_engine.services.policy_engine.engine.feeds.db import get_all_feeds
 from anchore_engine.services.policy_engine.engine.feeds.feeds import (
     DataFeed,
     GrypeDBFeed,
 )
-from anchore_engine.services.policy_engine.engine.feeds.sync import (
-    DataFeeds,
-    SourceFeeds,
-)
 from anchore_engine.subsys import logger
 
 GRYPE_DB_FEED_NAME = GrypeDBFeed.__feed_name__
+
+
+class MetadataSyncUtils:
+    @staticmethod
+    def _pivot_and_filter_feeds_by_config(
+        to_sync: list, source_found: list, db_found: list
+    ) -> Dict[str, FeedMetadata]:
+        """
+        Filters FeedMetadata records to only include those that are configured
+
+        :param to_sync: list of feed names requested to be synced
+        :param source_found: list of feed names available as returned by the upstream source
+        :param db_found: list of db records that were updated as result of upstream metadata sync (this is to handle db update failures)
+        :return: dict of feed names to FeedMetadata records
+        :rtype: Dict[str, FeedMetadata]
+        """
+        available = set(to_sync).intersection(set(source_found))
+        return {x.name: x for x in db_found if x.name in available}
+
+    @staticmethod
+    def _sync_feed_metadata(
+        db: Session,
+        feed_api_record: Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]],
+        db_feeds: Dict[str, FeedMetadata],
+        operation_id: Optional[str] = None,
+    ) -> Dict[str, FeedMetadata]:
+        """
+        Add FeedMetadata records to DB if they don't already exist
+
+        :param db: database session
+        :type db: Session
+        :param feed_api_record: data from API client
+        :type feed_api_record: Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        :param db_feeds: map of feed names to FeedMetadata
+        :type db_feeds: Dict[str, FeedMetadata]
+        :param operation_id: UUID4 hexadecimal string
+        :type operation_id: Optional[str]
+        :return: map of feed names to FeedMetadata that has been updated or created in the DB
+        :rtype: Dict[str, FeedMetadata]
+        """
+        api_feed = feed_api_record["meta"]
+        db_feed = db_feeds.get(api_feed.name)
+        # Do this instead of a db.merge() to ensure no timestamps are reset or overwritten
+        if not db_feed:
+            logger.debug(
+                "Adding new feed metadata record to db: {} (operation_id={})".format(
+                    api_feed.name, operation_id
+                )
+            )
+            db_feed = FeedMetadata(
+                name=api_feed.name,
+                description=api_feed.description,
+                access_tier=api_feed.access_tier,
+                enabled=True,
+            )
+            db.add(db_feed)
+            db.flush()
+            return {api_feed.name: db_feed}
+        else:
+            logger.debug(
+                "Feed metadata already in db: {} (operation_id={})".format(
+                    api_feed.name, operation_id
+                )
+            )
+            return db_feeds
+
+    @staticmethod
+    def _sync_feed_group_metadata(
+        db: Session,
+        feed_api_record: Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]],
+        db_feeds: Dict[str, FeedMetadata],
+        operation_id: Optional[str] = None,
+    ) -> None:
+        """
+        Add FeedGroupMetadata records to DB if they don't already exist
+
+        :param db: database session
+        :type db: Session
+        :param feed_api_record: data from API client
+        :type feed_api_record: Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        :param db_feeds: map of feed names to FeedMetadata tied to DB session
+        :type db_feeds: Dict[str, FeedMetadata]
+        :param operation_id: UUID4 hexadecimal string
+        :type operation_id: Optional[str]
+        """
+        api_feed = feed_api_record["meta"]
+        db_feed = db_feeds.get(api_feed.name)
+        # Check for any update
+        db_feed.description = api_feed.description
+        db_feed.access_tier = api_feed.access_tier
+
+        db_groups = {x.name: x for x in db_feed.groups}
+        for api_group in feed_api_record.get("groups", []):
+            db_group = db_groups.get(api_group.name)
+            # Do this instead of a db.merge() to ensure no timestamps are reset or overwritten
+            if not db_group:
+                logger.debug(
+                    "Adding new feed metadata record to db: {} (operation_id={})".format(
+                        api_group.name, operation_id
+                    )
+                )
+                db_group = FeedGroupMetadata(
+                    name=api_group.name,
+                    description=api_group.description,
+                    access_tier=api_group.access_tier,
+                    feed=db_feed,
+                    enabled=True,
+                )
+                db_group.last_sync = None
+                db.add(db_group)
+            else:
+                logger.debug(
+                    "Feed group metadata already in db: {} (operation_id={})".format(
+                        api_group.name, operation_id
+                    )
+                )
+
+            db_group.access_tier = api_group.access_tier
+            db_group.description = api_group.description
+
+    @staticmethod
+    def sync_metadata(
+        source_feeds: Dict[
+            str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        ],
+        to_sync: List[str] = None,
+        operation_id: Optional[str] = None,
+        groups: bool = True,
+    ) -> Tuple[Dict[str, FeedMetadata], List[Tuple[str, Union[str, BaseException]]]]:
+        """
+        Get metadata from source and sync db metadata records to that (e.g. add any new groups or feeds)
+        Executes as a unit-of-work for db, so will commit result and returns the records found on upstream source.
+
+        If a record exists in db but was not found upstream, it is not returned
+
+        :param source_feeds: mapping containing FeedAPIRecord and FeedAPIGroupRecord
+        :type source_feeds: Dict[str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]]
+        :param to_sync: list of string feed names to sync metadata on
+        :type to_sync: List[str]
+        :param operation_id: UUID4 hexadecimal string
+        :type operation_id: Optional[str]
+        :param groups: whether or not to sync group metadata (defaults to True, which will sync group metadata)
+        :type groups: bool
+        :return: tuple, first element: dict of names mapped to db records post-sync only including records successfully updated by upstream, second element is a list of tuples where each tuple is (failed_feed_name, error_obj)
+        :rtype: Tuple[Dict[str, FeedMetadata], List[Tuple[str, Union[str, BaseException]]]
+        """
+
+        if not to_sync:
+            return {}, []
+
+        db = get_session()
+        try:
+            logger.info(
+                "Syncing feed and group metadata from upstream source (operation_id={})".format(
+                    operation_id
+                )
+            )
+            failed = []
+            db_feeds = MetadataSyncUtils._pivot_and_filter_feeds_by_config(
+                to_sync, list(source_feeds.keys()), get_all_feeds(db)
+            )
+
+            for feed_name, feed_api_record in source_feeds.items():
+                try:
+                    logger.info(
+                        "Syncing metadata for feed: {} (operation_id={})".format(
+                            feed_name, operation_id
+                        )
+                    )
+                    feed_metadata_map = MetadataSyncUtils._sync_feed_metadata(
+                        db, feed_api_record, db_feeds, operation_id
+                    )
+                    if groups:
+                        MetadataSyncUtils._sync_feed_group_metadata(
+                            db, feed_api_record, feed_metadata_map, operation_id
+                        )
+                except Exception as e:
+                    logger.exception("Error syncing feed {}".format(feed_name))
+                    logger.warn(
+                        "Could not sync metadata for feed: {} (operation_id={})".format(
+                            feed_name, operation_id
+                        )
+                    )
+                    failed.append((feed_name, e))
+                finally:
+                    db.flush()
+
+            # Reload
+            db_feeds = MetadataSyncUtils._pivot_and_filter_feeds_by_config(
+                to_sync, list(source_feeds.keys()), get_all_feeds(db)
+            )
+
+            db.commit()
+            logger.info(
+                "Metadata sync from feeds upstream source complete (operation_id={})".format(
+                    operation_id
+                )
+            )
+            return db_feeds, failed
+        except Exception as e:
+            logger.error(
+                "Rolling back feed metadata update due to error: {} (operation_id={})".format(
+                    e, operation_id
+                )
+            )
+            db.rollback()
+            raise
 
 
 class SyncUtilProvider(ABC):
@@ -82,23 +290,29 @@ class SyncUtilProvider(ABC):
 
     @abstractmethod
     def sync_metadata(
-        self, source_feeds: SourceFeeds, operation_id: Optional[str]
+        self,
+        source_feeds: Dict[
+            str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        ],
+        operation_id: Optional[str],
     ) -> Tuple[Dict[str, FeedMetadata], List[Tuple[str, Union[str, BaseException]]]]:
         """
-        Wraps DataFeeds.sync_metadata so that it may be called with arguments appropriate for the provider.
+        Wraps MetadataSyncUtils.sync_metadata so that it may be called with arguments appropriate for the provider.
 
         :param source_feeds: mapping containing FeedAPIRecord and FeedAPIGroupRecord
-        :type source_feeds: SourceFeeds
+        :type source_feeds: Dict[str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]]
         :param operation_id: UUID4 hexadecimal string
         :type operation_id: Optional[str]
-        :return: response of DataFeeds.sync_metadata()
+        :return: response of MetadataSyncUtils.sync_metadata()
         :rtype: Tuple[Dict[str, FeedMetadata], List[Tuple[str, Union[str, BaseException]]]]
         """
         ...
 
     @staticmethod
     def get_groups_to_download(
-        source_feeds: SourceFeeds,
+        source_feeds: Dict[
+            str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        ],
         feeds_to_sync: List[DataFeed],
         operation_id: str,
     ) -> List[FeedGroupMetadata]:
@@ -106,7 +320,7 @@ class SyncUtilProvider(ABC):
         Returns a list of FeedGroupMetadata for each feed group to download.
 
         :param source_feeds: mapping containing FeedAPIRecord and FeedAPIGroupRecord
-        :type source_feeds: SourceFeeds
+        :type source_feeds: Dict[str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]]
         :param feeds_to_sync: ordered list of DataFeed(s) to sync
         :type feeds_to_sync: List[DataFeed]
         :param operation_id: UUID4 hexadecimal string
@@ -149,24 +363,30 @@ class LegacySyncUtilProvider(SyncUtilProvider):
         return get_feeds_client(sync_config)
 
     def sync_metadata(
-        self, source_feeds: SourceFeeds, operation_id: Optional[str]
+        self,
+        source_feeds: Dict[
+            str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        ],
+        operation_id: Optional[str],
     ) -> Tuple[Dict[str, FeedMetadata], List[Tuple[str, Union[str, BaseException]]]]:
         """
-        Wraps DataFeeds.sync_metadata so that it may be called with arguments appropriate for the provider.
+        Wraps MetadataSyncUtils.sync_metadata so that it may be called with arguments appropriate for the provider.
         In this case, we want to make sure that syncing FeedGroupMetadata is enabled for the legacy feeds.
 
         :param source_feeds: mapping containing FeedAPIRecord and FeedAPIGroupRecord
-        :type source_feeds: SourceFeeds
+        :type source_feeds: Dict[str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]]
         :param operation_id: UUID4 hexadecimal string
         :type operation_id: Optional[str]
-        :return: response of DataFeeds.sync_metadata()
+        :return: response of MetadataSyncUtils.sync_metadata()
         :rtype: Tuple[Dict[str, FeedMetadata], List[Tuple[str, Union[str, BaseException]]]]
         """
-        return DataFeeds.sync_metadata(source_feeds, self.to_sync, operation_id)
+        return MetadataSyncUtils.sync_metadata(source_feeds, self.to_sync, operation_id)
 
     @staticmethod
     def get_groups_to_download(
-        source_feeds: SourceFeeds,
+        source_feeds: Dict[
+            str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        ],
         feeds_to_sync: List[DataFeed],
         operation_id: str,
     ) -> List[FeedGroupMetadata]:
@@ -175,7 +395,7 @@ class LegacySyncUtilProvider(SyncUtilProvider):
         enabled == True.
 
         :param source_feeds: mapping containing FeedAPIRecord and FeedAPIGroupRecord
-        :type source_feeds: SourceFeeds
+        :type source_feeds: Dict[str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]]
         :param feeds_to_sync: ordered list of DataFeed(s) to sync
         :type feeds_to_sync: List[DataFeed]
         :param operation_id: UUID4 hexadecimal string
@@ -248,26 +468,32 @@ class GrypeDBSyncUtilProvider(SyncUtilProvider):
         return get_grype_db_client(grype_db_sync_config)
 
     def sync_metadata(
-        self, source_feeds: SourceFeeds, operation_id: Optional[str]
+        self,
+        source_feeds: Dict[
+            str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        ],
+        operation_id: Optional[str],
     ) -> Tuple[Dict[str, FeedMetadata], List[Tuple[str, Union[str, BaseException]]]]:
         """
-        Wraps DataFeeds.sync_metadata so that it may be called with arguments appropriate for the provider.
+        Wraps MetadataSyncUtils.sync_metadata so that it may be called with arguments appropriate for the provider.
         In this case, we want to make sure that syncing FeedGroupMetadata is disabled for grypedb feed.
 
         :param source_feeds: mapping containing FeedAPIRecord and FeedAPIGroupRecord
-        :type source_feeds: SourceFeeds
+        :type source_feeds: Dict[str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]]
         :param operation_id: UUID4 hexadecimal string
         :type operation_id: Optional[str]
-        :return: response of DataFeeds.sync_metadata()
+        :return: response of MetadataSyncUtils.sync_metadata()
         :rtype: Tuple[Dict[str, FeedMetadata], List[Tuple[str, Union[str, BaseException]]]]
         """
-        return DataFeeds.sync_metadata(
+        return MetadataSyncUtils.sync_metadata(
             source_feeds, self.to_sync, operation_id, groups=False
         )
 
     @staticmethod
     def get_groups_to_download(
-        source_feeds: SourceFeeds,
+        source_feeds: Dict[
+            str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]
+        ],
         feeds_to_sync: List[DataFeed],
         operation_id: str,
     ) -> List[FeedGroupMetadata]:
@@ -277,7 +503,7 @@ class GrypeDBSyncUtilProvider(SyncUtilProvider):
         Uses FeedMetadata from feeds_to_sync. Expects only one record is present for grypedb.
 
         :param source_feeds: mapping containing FeedAPIRecord and FeedAPIGroupRecord
-        :type source_feeds: SourceFeeds
+        :type source_feeds: Dict[str, Dict[str, Union[FeedAPIRecord, List[FeedAPIGroupRecord]]]]
         :param feeds_to_sync: ordered list of DataFeed(s) to sync
         :type feeds_to_sync: List[DataFeed]
         :param operation_id: UUID4 hexadecimal string
